global refactoring.
the whole project is refactored.
tests are needed!
