+++ app/config/adapters_config.json
