+++ app/static/adapters.js
