+++ app/static/js/adapters.js
