+++ app/templates/adapters.html
